drivers: firmware: qcom: Enable download mode
In order to aid post-mortem debugging the Qualcomm platforms provide a
"memory download mode", where the boot loader will provide an interface
for custom tools to "download" the content of RAM to a host machine.

The mode is triggered by writing a magic value somewhere in RAM, that is
read in the boot code path after a warm-restart. Two mechanism for
setting this magic value are supported in modern platforms; a direct SCM
call to enable the mode or through a secure io write of a magic value.

In order for a normal reboot not to trigger "download mode" the magic
must be cleared during a clean reboot.

Download mode has to be enabled by including qcom_scm.download_mode=1 on
the command line.

Change-Id: I622667e5055def90dea0e1799ec7fc3a2453d7bd
Signed-off-by: Yossi Mansharoff <yossim@codeaurora.org>
Signed-off-by: Ramon Fried <rfried@codeaurora.org>